refactor(schemas): move Zod-only schemas to _zod/ subdirectory
Separates Zod validators from Drizzle tables for clarity. Moved:
- schemas/user.ts → schemas/_zod/user.ts
- schemas/rbac.ts → schemas/_zod/rbac.ts
- schemas/validation.ts → schemas/_zod/validation.ts

All 9 direct importers updated to the new path. Top-level barrel
re-exports the entire _zod surface so existing consumers keep working.
