+++ packages/nextly/src/schemas/_zod/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Zod-only schemas (validators, types) — kept separate from Drizzle tables for clarity.
+ *
+ * Import via `import { CreatePermissionSchema } from "nextly/schemas/_zod/rbac"` or
+ * via the top-level `nextly/schemas` re-export for backward compatibility.
+ *
+ * @module schemas/_zod
+ * @since v0.0.3-alpha (Plan A — schemas consolidation)
+ */
+
+export * from "./user";
+export * from "./rbac";
+export * from "./validation";
